+++ modules/common/adapters/adapter_gflags.cc
@@ -117,6 +117,12 @@ DEFINE_string(system_status_topic, "/apollo/monitor/system_status",
 DEFINE_string(static_info_topic, "/apollo/monitor/static_info",
               "Static info topic name");
 DEFINE_string(mobileye_topic, "/apollo/sensor/mobileye", "mobileye topic name");
+DEFINE_string(smartereye_obstacles_topic, "/apollo/sensor/smartereye/obstacles",
+              "smartereye obstacles topic name");
+DEFINE_string(smartereye_lanemark_topic, "/apollo/sensor/smartereye/lanemark",
+              "smartereye lanemark topic name");
+DEFINE_string(smartereye_image_topic, "/apollo/sensor/smartereye/image",
+              "smartereye image topic name");
 DEFINE_string(delphi_esr_topic, "/apollo/sensor/delphi_esr",
               "delphi esr radar topic name");
 DEFINE_string(conti_radar_topic, "/apollo/sensor/conti_radar",
