+++ modules/common/adapters/adapter_gflags.h
@@ -64,6 +64,9 @@ DECLARE_string(gnss_status_topic);
 DECLARE_string(system_status_topic);
 DECLARE_string(static_info_topic);
 DECLARE_string(mobileye_topic);
+DECLARE_string(smartereye_obstacles_topic);
+DECLARE_string(smartereye_lanemark_topic);
+DECLARE_string(smartereye_image_topic);
 DECLARE_string(delphi_esr_topic);
 DECLARE_string(conti_radar_topic);
 DECLARE_string(racobit_radar_topic);
