+++ modules/drivers/proto/BUILD
@@ -15,6 +15,7 @@ proto_library(
         "conti_radar.proto",
         "delphi_esr.proto",
         "mobileye.proto",
+        "smartereye.proto",
         "pointcloud.proto",
         "racobit_radar.proto",
         "radar.proto",
