+++ modules/third_party_perception/BUILD
@@ -4,47 +4,51 @@ package(default_visibility = ["//visibility:public"])
 
 cc_library(
     name = "third_party_perception_lib",
-    srcs = ["third_party_perception.cc"],
-    hdrs = ["third_party_perception.h"],
+    srcs = ["third_party_perception_base.cc"],
+    hdrs = ["third_party_perception_base.h"],
     deps = [
-        ":third_party_perception_conversion",
-        ":third_party_perception_filter",
-        ":third_party_perception_fusion",
         "//cyber",
         "//modules/common/adapters:adapter_gflags",
         "//modules/common/util:message_util",
+        "//modules/third_party_perception/tools:third_party_perception_filter",
+        "//modules/third_party_perception/tools:third_party_perception_fusion",
+        "//modules/third_party_perception/tools:third_party_perception_radar_conversion",
+        "//modules/third_party_perception/tools:third_party_perception_smartereye_conversion",
+        "//modules/third_party_perception/tools:third_party_perception_mobileye_conversion",               
     ],
 )
 
 cc_library(
-    name = "third_party_perception_conversion",
-    srcs = ["conversion.cc"],
-    hdrs = ["conversion.h"],
+    name = "third_party_perception_mobileye_lib",
+    srcs = ["third_party_perception_mobileye.cc"],
+    hdrs = ["third_party_perception_mobileye.h"],
     deps = [
-        "//modules/canbus/proto:canbus_proto",
-        "//modules/drivers/proto:sensor_proto",
-        "//modules/third_party_perception/common:third_party_perception_util",
-        "//modules/third_party_perception/proto:third_party_perception_proto",
+        ":third_party_perception_lib",
+        "//cyber",
+        "//modules/common/adapters:adapter_gflags",
+        "//modules/common/util:message_util",
+        "//modules/third_party_perception/tools:third_party_perception_filter",
+        "//modules/third_party_perception/tools:third_party_perception_fusion",
+        "//modules/third_party_perception/tools:third_party_perception_radar_conversion",
+        "//modules/third_party_perception/tools:third_party_perception_smartereye_conversion",
+        "//modules/third_party_perception/tools:third_party_perception_mobileye_conversion",
     ],
 )
 
 cc_library(
-    name = "third_party_perception_fusion",
-    srcs = ["fusion.cc"],
-    hdrs = ["fusion.h"],
+    name = "third_party_perception_smartereye_lib",
+    srcs = ["third_party_perception_smartereye.cc"],
+    hdrs = ["third_party_perception_smartereye.h"],
     deps = [
-        "//modules/common/math:geometry",
-        "//modules/perception/proto:perception_proto",
-    ],
-)
-
-cc_library(
-    name = "third_party_perception_filter",
-    srcs = ["filter.cc"],
-    hdrs = ["filter.h"],
-    deps = [
-        "//modules/third_party_perception/common:third_party_perception_util",
-        "//modules/third_party_perception/proto:third_party_perception_proto",
+        ":third_party_perception_lib",
+        "//cyber",
+        "//modules/common/adapters:adapter_gflags",
+        "//modules/common/util:message_util",
+        "//modules/third_party_perception/tools:third_party_perception_filter",
+        "//modules/third_party_perception/tools:third_party_perception_fusion",
+        "//modules/third_party_perception/tools:third_party_perception_radar_conversion",
+        "//modules/third_party_perception/tools:third_party_perception_smartereye_conversion",
+        "//modules/third_party_perception/tools:third_party_perception_mobileye_conversion",
     ],
 )
 
@@ -55,6 +59,8 @@ cc_library(
     copts = ['-DMODULE_NAME=\\"third_party_perception\\"'],
     deps = [
         ":third_party_perception_lib",
+        ":third_party_perception_mobileye_lib",
+        ":third_party_perception_smartereye_lib",
     ],
 )
