+++ modules/third_party_perception/conf/third_party_perception.conf
@@ -10,6 +10,6 @@
 --default_unknown_width=2.0
 --default_bike_width=2.0
 --default_bike_length=5.0
---enable_radar=true
+--enable_radar=false
 --movable_heading_threshold=0.8
---use_navigation_mode
+--nouse_navigation_mode
