+++ modules/third_party_perception/conf/third_party_perception_component.pb.txt
@@ -1,0 +1,1 @@
+device_names: "ThirdPartyPerceptionSmartereye"
