+++ modules/third_party_perception/third_party_perception_component.cc
@@ -16,7 +16,11 @@
 
 #include "modules/third_party_perception/third_party_perception_component.h"
 
+#include "boost/algorithm/string.hpp"
+#include "boost/format.hpp"
+
 #include "modules/common/adapters/adapter_gflags.h"
+#include "modules/third_party_perception/proto/third_party_perception_component.pb.h"
 
 DECLARE_string(flagfile);
 
@@ -24,58 +28,45 @@ namespace apollo {
 namespace third_party_perception {
 
 bool ThirdPartyPerceptionComponent::Init() {
-  if (!perception_.Init().ok()) {
+  apollo::third_party_perception::ThirdPartyPerceptionDevice
+          third_party_perception_param;
+  if (!GetProtoConfig(&third_party_perception_param)) {
+    AINFO << "load third party perception param failed";
+    return false;
+  }
+
+  std::string device_names_str = third_party_perception_param.device_names();
+  std::vector<std::string> device_names;
+
+  boost::algorithm::split(device_names, device_names_str,
+                          boost::algorithm::is_any_of(","));
+  if (device_names.size() != 1) {
+    AERROR << "Now third_party_perception only support one camera device";
+    return false;
+  } else {
+    if (device_names.at(0) == "ThirdPartyPerceptionSmartereye") {
+      perception_ = std::make_shared<ThirdPartyPerceptionSmartereye>(
+                    node_.get());
+    } else if (device_names.at(0) == "ThirdPartyPerceptionMobileye") {
+      perception_ = std::make_shared<ThirdPartyPerceptionMobileye>(node_.get());
+    } else {
+      perception_ = std::make_shared<ThirdPartyPerception>(node_.get());
+    }
+  }
+
+  if (!perception_->Init().ok()) {
     return false;
   }
 
   writer_ = node_->CreateWriter<apollo::perception::PerceptionObstacles>(
       FLAGS_perception_obstacle_topic);
 
-  std::shared_ptr<Reader<apollo::drivers::Mobileye>> mobileye_reader_ = nullptr;
-  std::shared_ptr<Reader<apollo::drivers::DelphiESR>> delphi_esr_reader_ =
-      nullptr;
-  std::shared_ptr<Reader<apollo::drivers::ContiRadar>> conti_radar_reader_ =
-      nullptr;
-
-  mobileye_reader_ = node_->CreateReader<apollo::drivers::Mobileye>(
-      FLAGS_mobileye_topic,
-      [this](const std::shared_ptr<apollo::drivers::Mobileye> &message) {
-        perception_.OnMobileye(*message.get());
-      });
-
-  delphi_esr_reader_ = node_->CreateReader<apollo::drivers::DelphiESR>(
-      FLAGS_delphi_esr_topic,
-      [this](const std::shared_ptr<apollo::drivers::DelphiESR> &message) {
-        perception_.OnDelphiESR(*message.get());
-      });
-
-  conti_radar_reader_ = node_->CreateReader<apollo::drivers::ContiRadar>(
-      FLAGS_conti_radar_topic,
-      [this](const std::shared_ptr<apollo::drivers::ContiRadar> &message) {
-        perception_.OnContiRadar(*message.get());
-      });
-
-  localization_reader_ =
-      node_->CreateReader<apollo::localization::LocalizationEstimate>(
-          FLAGS_localization_topic,
-          [this](
-              const std::shared_ptr<apollo::localization::LocalizationEstimate>
-                  &localization) {
-            perception_.OnLocalization(*localization.get());
-          });
-
-  chassis_reader_ = node_->CreateReader<apollo::canbus::Chassis>(
-      FLAGS_chassis_topic,
-      [this](const std::shared_ptr<apollo::canbus::Chassis> &chassis) {
-        perception_.OnChassis(*chassis.get());
-      });
-
-  return perception_.Start().ok();
+  return perception_->Start().ok();
 }
 
 bool ThirdPartyPerceptionComponent::Proc() {
   auto response = std::make_shared<apollo::perception::PerceptionObstacles>();
-  if (!perception_.Process(response.get())) {
+  if (!perception_->Process(response.get())) {
     return false;
   }
   writer_->Write(response);
