+++ modules/third_party_perception/third_party_perception_component.h
@@ -17,11 +17,15 @@
 #pragma once
 
 #include <memory>
+#include <string>
+#include <vector>
 
 #include "cyber/component/component.h"
 #include "cyber/component/timer_component.h"
-
-#include "modules/third_party_perception/third_party_perception.h"
+#include "modules/drivers/proto/sensor_image.pb.h"
+#include "modules/third_party_perception/third_party_perception_base.h"
+#include "modules/third_party_perception/third_party_perception_smartereye.h"
+#include "modules/third_party_perception/third_party_perception_mobileye.h"
 
 namespace apollo {
 namespace third_party_perception {
@@ -37,21 +41,10 @@ class ThirdPartyPerceptionComponent final
   bool Proc() override;
 
  private:
-  std::shared_ptr<apollo::cyber::Reader<apollo::drivers::Mobileye>>
-      mobileye_reader_ = nullptr;
-  std::shared_ptr<apollo::cyber::Reader<apollo::drivers::DelphiESR>>
-      delphi_esr_reader_ = nullptr;
-  std::shared_ptr<apollo::cyber::Reader<apollo::drivers::ContiRadar>>
-      conti_radar_reader_ = nullptr;
-  std::shared_ptr<
-      apollo::cyber::Reader<apollo::localization::LocalizationEstimate>>
-      localization_reader_ = nullptr;
-  std::shared_ptr<apollo::cyber::Reader<apollo::canbus::Chassis>>
-      chassis_reader_ = nullptr;
   std::shared_ptr<
       apollo::cyber::Writer<apollo::perception::PerceptionObstacles>>
       writer_ = nullptr;
-  ThirdPartyPerception perception_;
+  std::shared_ptr<ThirdPartyPerception> perception_ = nullptr;
 };
 
 CYBER_REGISTER_COMPONENT(ThirdPartyPerceptionComponent)
