+++ modules/third_party_perception/third_party_perception_smartereye.h
@@ -1,0 +1,55 @@
+/******************************************************************************
+ * Copyright 2017 The Apollo Authors. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *****************************************************************************/
+
+/**
+ * @file
+ */
+
+#pragma once
+
+#include <memory>
+
+#include "modules/third_party_perception/third_party_perception_base.h"
+#include "modules/drivers/proto/smartereye.pb.h"
+
+/**
+ * @namespace apollo::third_party_perception
+ * @brief apollo::third_party_perception
+ */
+namespace apollo {
+namespace third_party_perception {
+
+class ThirdPartyPerceptionSmartereye : public ThirdPartyPerception {
+ public:
+  explicit ThirdPartyPerceptionSmartereye(apollo::cyber::Node* const node);
+  ThirdPartyPerceptionSmartereye() = default;
+  ~ThirdPartyPerceptionSmartereye() = default;
+  // Upon receiving smartereye data
+  void OnSmartereye(const apollo::drivers::SmartereyeObstacles& message);
+  void OnSmartereyeLanemark(const apollo::drivers::SmartereyeLanemark&);
+  bool Process(
+      apollo::perception::PerceptionObstacles* const response) override;
+ private:
+  apollo::perception::PerceptionObstacles eye_obstacles_;
+  apollo::drivers::SmartereyeLanemark smartereye_lanemark_;
+  std::shared_ptr<apollo::cyber::Reader<apollo::drivers::SmartereyeObstacles>>
+      smartereye_obstacles_reader_ = nullptr;
+  std::shared_ptr<apollo::cyber::Reader<apollo::drivers::SmartereyeLanemark>>
+      smartereye_lanemark_reader_ = nullptr;
+};
+
+}  // namespace third_party_perception
+}  // namespace apollo
