+++ modules/third_party_perception/tools/conversion_mobileye.h
@@ -21,39 +21,22 @@
 #pragma once
 
 #include "modules/canbus/proto/chassis.pb.h"
-#include "modules/drivers/proto/conti_radar.pb.h"
-#include "modules/drivers/proto/delphi_esr.pb.h"
 #include "modules/drivers/proto/mobileye.pb.h"
 #include "modules/localization/proto/localization.pb.h"
-#include "modules/third_party_perception/proto/radar_obstacle.pb.h"
 
 /**
- * @namespace apollo::third_party_perception::conversion
+ * @namespace apollo::third_party_perception::conversion_mobileye
  * @brief apollo::third_party_perception
  */
 namespace apollo {
 namespace third_party_perception {
-namespace conversion {
+namespace conversion_mobileye {
 
 apollo::perception::PerceptionObstacles MobileyeToPerceptionObstacles(
     const apollo::drivers::Mobileye& mobileye,
     const apollo::localization::LocalizationEstimate& localization,
     const apollo::canbus::Chassis& chassis);
 
-RadarObstacles DelphiToRadarObstacles(
-    const apollo::drivers::DelphiESR& delphi_esr,
-    const apollo::localization::LocalizationEstimate& localization,
-    const RadarObstacles& last_radar_obstacles);
-
-RadarObstacles ContiToRadarObstacles(
-    const apollo::drivers::ContiRadar& conti_radar,
-    const apollo::localization::LocalizationEstimate& localization,
-    const RadarObstacles& last_radar_obstacles,
-    const apollo::canbus::Chassis& chassis);
-
-apollo::perception::PerceptionObstacles RadarObstaclesToPerceptionObstacles(
-    const RadarObstacles& radar_obstacles);
-
-}  // namespace conversion
+}  // namespace conversion_mobileye
 }  // namespace third_party_perception
 }  // namespace apollo
