+++ modules/third_party_perception/tools/filter.cc
@@ -18,7 +18,7 @@
  * @file
  */
 
-#include "modules/third_party_perception/filter.h"
+#include "modules/third_party_perception/tools/filter.h"
 
 #include "modules/third_party_perception/common/third_party_perception_gflags.h"
 #include "modules/third_party_perception/common/third_party_perception_util.h"
