+++ scripts/third_party_perception.sh
@@ -25,4 +25,6 @@ source "$DIR/apollo_base.sh"
 
 # run function from apollo_base.sh
 # run command_name module_name
-run third_party_perception "$@" --use_navigation_mode
+# run third_party_perception "$@"
+cyber_launch stop modules/third_party_perception/launch/third_party_perception.launch
+cyber_launch start modules/third_party_perception/launch/third_party_perception.launch
